+++ wasock/__init__.py
@@ -2,5 +2,6 @@ from .whatsappsocket import WhatsAppSocket
 from .qrcode import QRCode
 from .connection import Connection
 from .message import Message
+from .browser import Browser
 
-__version__ = "WAsock @0.5.1-Beta"
+__version__ = "WAsock @0.5.2"